divided up module
